// tests/fixtures/petstore-paths.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  '/pets': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * List pets
     * @description Returns all pets in the store
     */
    get: operations['listPets'];
    put?: never;
    /** Create a pet */
    post: operations['createPet'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/pets/{petId}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get pet by ID */
    get: operations['getPetById'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    Pet: {
      id: number;
      name: string;
      tag?: string;
    };
    Error: {
      code: number;
      message: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
  listPets: {
    parameters: {
      query?: {
        /** @description How many items to return */
        limit?: number;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description A paged array of pets */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['Pet'][];
        };
      };
      /** @description unexpected error */
      default: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['Error'];
        };
      };
    };
  };
  createPet: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** @description Pet to add */
    requestBody: {
      content: {
        'application/json': components['schemas']['Pet'];
      };
    };
    responses: {
      /** @description Created */
      201: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['Pet'];
        };
      };
    };
  };
  getPetById: {
    parameters: {
      query?: never;
      header?: never;
      path: {
        /** @description ID of pet to fetch */
        petId: string;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description Expected response */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['Pet'];
        };
      };
    };
  };
}